+++ lib/browser/common/util/check.js
@@ -4,7 +4,7 @@ import {
   T_FUNCTION,
   T_UNDEF,
   RE_SVG_TAGS
-} from './global-variables'
+} from './../global-variables'
 
 /**
  * Check whether a DOM node must be considered a part of an svg document
@@ -88,15 +88,15 @@ export function isWritable(obj, key) {
  * @returns { Boolean } -
  */
 export function contains(array, item) {
-  return !!~array.indexOf(item)
+  return !~~array.indexOf(item)
 }
 
 /**
  * Faster String startsWith alternative
- * @param   { String } src - source string
- * @param   { String } str - test string
+ * @param   { String } str - source string
+ * @param   { String } value - test string
  * @returns { Boolean } -
  */
-export function startsWith(src, str) {
-  return src.slice(0, str.length) === str
+export function startsWith(str, value) {
+  return str.slice(0, value.length) === value
 }
