+++ lib/browser/compiler/index.js
@@ -1,8 +1,15 @@
 import observable from 'riot-observable'
 import compiler from 'riot-compiler'
 import { T_STRING } from './../common/global-variables'
-import { isFunction, isArray, isObject } from './../common/check'
-import { extend, $$, getAttr, mkEl } from './../common/util'
+import {
+  isFunction,
+  isArray,
+  isObject,
+  extend,
+  $$,
+  getAttr,
+  mkEl
+} from './../common/util'
 
 var
   promise,    // emits the 'ready' event and runs the first callback
