+++ lib/browser/tag/core.js
@@ -12,10 +12,7 @@ import {
   isString,
   isUndefined,
   isObject,
-  isFunction
-} from './../common/check'
-
-import {
+  isFunction,
   setAttr,
   getAttr,
   $$,
