+++ lib/browser/tag/each.js
@@ -7,10 +7,7 @@ import {
 } from './../common/global-variables'
 
 import {
-  isArray
-} from './../common/check'
-
-import {
+  isArray,
   remAttr,
   getAttr,
   getTagName,
