+++ lib/browser/tag/mkdom.js
@@ -5,10 +5,7 @@ import {
 } from './../common/global-variables'
 
 import {
-  isSVGTag
-} from './../common/check'
-
-import {
+  isSVGTag,
   mkEl,
   setInnerHTML,
   $
