+++ lib/browser/tag/named.js
@@ -1,10 +1,7 @@
 import { tmpl } from 'riot-tmpl'
 
 import {
-  isBlank
-} from './../common/check'
-
-import {
+  isBlank,
   getImmediateCustomParentTag,
   arrayishAdd,
   setAttr,
