+++ lib/browser/tag/tag.js
@@ -21,10 +21,7 @@ import {
   isFunction,
   isObject,
   isWritable,
-  contains
-} from './../common/check'
-
-import {
+  contains,
   inherit,
   cleanUpData,
   defineProperty,
