+++ lib/browser/tag/update.js
@@ -14,10 +14,7 @@ import {
   startsWith,
   isFunction,
   isArray,
-  isWritable
-} from './../common/check'
-
-import {
+  isWritable,
   each,
   remAttr,
   setAttr,
